# sentinel/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Defentrax API
  description: |
    Control-plane REST API for Defentrax (security monitoring / lightweight SIEM).

    This specification documents **implemented** routes only (inventory from
    `sentinel/api/internal/server/server.go`). It does **not** invent endpoints.

    **Not exposed as REST (yet):**
    - No standalone `GET /api/v1/agents` — agents appear nested under `GET /api/v1/servers/{id}`.
    - No `/api/v1/metrics` — use `GET /api/v1/dashboard/stats`.

    **Page permissions:** panel routes require a `pages:*` key in addition to the
    data permission named on each operation. `pages:dashboard`, `pages:alerts`,
    `pages:alert_detail`, `pages:events`, `pages:servers`, `pages:server_detail`,
    `pages:rules`, `pages:notifications`, `pages:team`, `pages:roles`, `pages:audit`.
    `SUPER_ADMIN` and `ADMIN` bypass these checks. Plugins and API keys have no page key.

    Interactive UI: `GET /api/v1/docs` (Swagger UI). Raw spec: `GET /api/v1/openapi.yaml`.
  version: 0.2.0
  contact:
    name: Defentrax
servers:
  - url: http://localhost:8080
    description: Local default (API_PORT / PORT)
...
paths:
...
  /api/v1/servers:
    get:
      tags: [Servers]
      operationId: listServers
      summary: List servers
      description: "Requires `servers:read` and `pages:servers`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      responses:
        "200":
          description: Server list
          content:
            application/json:
              schema:
                type: object
                required: [servers]
                properties:
                  servers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Server"
    post:
      tags: [Servers]
      operationId: createServer
      summary: Create server
      description: "Requires `servers:write` and `pages:servers`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                hostname:
                  type: string
                description:
                  type: string
                environment:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
  /api/v1/servers/{id}:
    get:
      tags: [Servers, Agents]
      operationId: getServer
      summary: Get server with nested agents
      description: |
        Requires `servers:read` and `pages:server_detail`.
        Agents are returned nested here; there is no standalone agents list endpoint.
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      responses:
        "200":
          description: Server detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerDetail"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/servers/{id}/enrollment-tokens:
    post:
      tags: [Servers, Agents]
      operationId: createEnrollmentToken
      summary: Create one-time enrollment token
      description: |
        Requires `servers:write` and `pages:server_detail`.
        Returns a one-time `senr_…` token (shown once). Default TTL 60 minutes (max 24h).
      security:
        - SessionCookie: []
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/PathUUID"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                ttl_minutes:
                  type: integer
                  minimum: 1
                  maximum: 1440
                  default: 60
      responses:
        "201":
          description: Token created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentToken"
...
  /api/v1/rules:
    get:
      tags: [Rules]
      operationId: listRules
      summary: List detection rules
      description: "Requires `rules:read` and `pages:rules`."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      responses:
        "200":
          description: Rules
          content:
            application/json:
              schema:
                type: object
                required: [rules]
                properties:
                  rules:
                    type: array
                    items:
                      $ref: "#/components/schemas/Rule"
    post:
      tags: [Rules]
      operationId: createRule
      summary: Create a custom detection rule
      description: "Requires `rules:write` and `pages:rules`. The rule is stored and loaded into the detection engine. Names and alert text must stay possibility-oriented. At least one match field is required."
      security:
        - SessionCookie: []
        - BearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description, severity]
              properties:
                name:
                  type: string
                description:
                  type: string
                severity:
                  type: string
                  enum: [info, low, medium, high, critical]
                source:
                  type: string
                category:
                  type: string
                event_type:
                  type: string
                message_contains:
                  type: string
                fields:
                  type: object
                  additionalProperties:
                    type: string
                threshold_count:
                  type: integer
                window_seconds:
                  type: integer
                group_by:
                  type: array
                  items:
                    type: string
                action_title:
                  type: string
                action_description:
                  type: string
      responses:
        "201":
          description: Created rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sentinel_session
      description: "Server-side session cookie (name configurable via `SESSION_COOKIE_NAME`; default `sentinel_session`)."
    BearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: sent_
      description: "Personal API key (`Authorization: Bearer sent_…`)."
    BearerAgentToken:
      type: http
      scheme: bearer
      bearerFormat: sagt_
      description: "Agent token from enrollment (`Authorization: Bearer sagt_…`)."
  parameters:
    PathUUID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: unauthorized
        message:
          type: string
        request_id:
          type: string
...
    Server:
      type: object
      required: [id, name, hostname, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        hostname:
          type: string
        created_at:
          type: string
          format: date-time
    Agent:
      type: object
      required: [id, server_id, name, status, agent_version]
      properties:
        id:
          type: string
          format: uuid
        server_id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        agent_version:
          type: string
        last_heartbeat_at:
          type: string
          format: date-time
    ServerDetail:
      allOf:
        - $ref: "#/components/schemas/Server"
        - type: object
          properties:
            description:
              type: string
            environment:
              type: string
            agents:
              type: array
              items:
                $ref: "#/components/schemas/Agent"
    EnrollmentToken:
      type: object
      required: [id, server_id, token_prefix, expires_at]
      properties:
        id:
          type: string
          format: uuid
        server_id:
          type: string
          format: uuid
        token_prefix:
          type: string
        token:
          type: string
          description: "Full `senr_…` token (create response only)"
        expires_at:
          type: string
          format: date-time
...
    Rule:
      type: object
      required: [id, rule_id, name, enabled, severity, version, definition, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        rule_id:
          type: string
          description: Stable YAML rule id
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        severity:
          type: string
        version:
          type: integer
        definition:
          description: Rule definition JSON
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...